requirements.txt
opencv-python
dlib
imutils
scipy
transformers
Pillow
numpy
torch
psutil
